refactor(test-integration): replace simulateClusterServiceCreate with setClusterServiceID
The simulateClusterServiceCreate step faked a real mock PostCluster call just
to stamp a ClusterServiceID onto a cluster doc. Now that DeriveClusterServiceID
can generate the same HREF deterministically (ocm.GenerateAROHCPClusterHREF),
node pools/external auths, and clusters can all use the lighter
setClusterServiceID + loadCosmos(SPC) pattern instead, without depending on
mock PostCluster history.

- DeriveClusterServiceID's cluster branch is now deterministic; drops the
  ClusterServiceMock/testName params it no longer needs.
- EnsureParentClusterServiceID moves into set_cluster_service_id.go and
  auto-provisions the parent cluster's ID via derive+set instead of a real
  mock create.
- Deletes SimulateBackendClusterServiceCreate, the simulateClusterServiceCreate
  and clusterServiceCompare step types, and the now-unused
  FindClusterHREF/GetMergedClusters mock helpers (equivalent tag/AFEC
  propagation coverage already exists in TestBuildCSCluster).
- Migrates the Cluster/experimental-features* artifact tests off
  simulateClusterServiceCreate+clusterServiceCompare onto loadCosmos(SPC)+
  setClusterServiceID, dropping the setClusterServiceID step for
  experimental-features specifically since cluster updates still
  synchronously call the real CS mock's GetCluster.
