--- test-integration/frontend/artifacts/FrontendCRUD/Cluster/experimental-features-no-afec/04-clusterServiceCompare-tags-ignored/expected-content.json
@@ -1,3 +1,0 @@
-{
-	"properties": {}
-}
