+++ test-integration/frontend/artifacts/FrontendCRUD/Cluster/experimental-features-no-afec/04-setClusterServiceID-cluster/00-key.json
@@ -1,5 +1,5 @@
 {
-	"_artifactStep": "simulateClusterServiceCreate",
+	"_artifactStep": "setClusterServiceID",
 	"id": "f94d3f60-1681-5b9f-9449-c48504f3cddb",
 	"resourceID": "/subscriptions/a1b2c3d4-0000-0000-0000-afec00000000/resourceGroups/resourceGroupName/providers/Microsoft.RedHatOpenShift/hcpOpenShiftClusters/no-afec-test"
 }
