--- test-integration/frontend/artifacts/FrontendCRUD/Cluster/experimental-features/03-simulateClusterServiceCreate-cluster/00-key.json
@@ -1,5 +1,0 @@
-{
-	"_artifactStep": "simulateClusterServiceCreate",
-	"id": "d189de99-ebfa-5c9f-846c-6656f3718661",
-	"resourceID": "/subscriptions/6b690bec-0c16-4ecb-8f67-781caf40bba7/resourceGroups/resourceGroupName/providers/Microsoft.RedHatOpenShift/hcpOpenShiftClusters/feature-flag-test"
-}
