--- test-integration/frontend/artifacts/FrontendCRUD/Cluster/experimental-features/04-clusterServiceCompare-create-propagated-flag/expected-content.json
@@ -1,6 +1,0 @@
-{
-	"properties": {
-		"hosted_cluster_single_replica": "true",
-		"hosted_cluster_size_override": "e2e_minimal"
-	}
-}
