+++ test-integration/frontend/artifacts/FrontendCRUD/Cluster/experimental-features/04-httpReplace-subscription-keep-afec/subscription.json
