--- test-integration/frontend/artifacts/FrontendCRUD/Cluster/experimental-features/07-clusterServiceCompare-update-cleared-flag/expected-content.json
@@ -1,3 +1,0 @@
-{
-	"properties": {}
-}
