+++ test-integration/utils/databasemutationhelpers/resource_crud_test_util.go
@@ -265,10 +265,6 @@ func NewUntypedStep(indexString, stepType, stepName string, testDir fs.FS, path 
 		return newCompleteOperationStep(stepID, stepDir)
 	case "setClusterServiceID":
 		return newSetClusterServiceIDStep(stepID, stepDir)
-	case "simulateClusterServiceCreate":
-		return newSimulateClusterServiceCreateStep(stepID, stepDir)
-	case "clusterServiceCompare":
-		return newClusterServiceCompareStep(stepID, stepDir)
 	case "migrateCosmos":
 		return newMigrateCosmosStep(stepID, stepDir)
 	case "kubernetesLoad":
@@ -362,11 +358,6 @@ func NewStep[InternalAPIType any, InternalAPITypePointer arm.CosmosMetadataAcces
 
 	case "setClusterServiceID":
 		return newSetClusterServiceIDStep(stepID, stepDir)
-	case "simulateClusterServiceCreate":
-		return newSimulateClusterServiceCreateStep(stepID, stepDir)
-
-	case "clusterServiceCompare":
-		return newClusterServiceCompareStep(stepID, stepDir)
 
 	case "migrateCosmos":
 		return newMigrateCosmosStep(stepID, stepDir)
